+++ cmd/cipdip/main.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	version = "0.2.7"
+	version = "0.2.8"
 	commit  = "unknown"
 	date    = "unknown"
 )